remove old version, replace by version 2 and delete version 2 .py
